cfpc_backend/requirements.txt
mysqlclient
Django
python-dotenv
djangorestframework
drf-spectacular
djangorestframework-simplejwt
djoser
json
regex